[fix] separeted implementation of api files and request files
